Rename `analytics_event.py` back to `event.py`
Avoids a breaking change for SDK users.
